+++ core/station/impl/src/models/request.rs
@@ -16,10 +16,7 @@ use crate::core::validation::{
     EnsureAccount, EnsureAddressBookEntry, EnsureIdExists, EnsureRequestPolicy, EnsureUser,
     EnsureUserGroup,
 };
-use crate::errors::{
-    EvaluateError, ExternalCanisterValidationError, RecordValidationError, RequestError,
-    ValidationError,
-};
+use crate::errors::{EvaluateError, RequestError, ValidationError};
 use crate::repositories::USER_REPOSITORY;
 use candid::{CandidType, Deserialize};
 use orbit_essentials::repository::Repository;
@@ -230,37 +227,6 @@ fn validate_request_operation_foreign_keys(
         }
     }
     Ok(())
-}
-
-impl From<RecordValidationError> for RequestError {
-    fn from(err: RecordValidationError) -> RequestError {
-        match err {
-            RecordValidationError::NotFound { id, model_name } => RequestError::ValidationError {
-                info: format!("Invalid user specifier: {} {} not found", model_name, id),
-            },
-        }
-    }
-}
-
-impl From<ExternalCanisterValidationError> for RequestError {
-    fn from(err: ExternalCanisterValidationError) -> RequestError {
-        match err {
-            ExternalCanisterValidationError::InvalidExternalCanister { principal } => {
-                RequestError::ValidationError {
-                    info: format!("Invalid external canister {}", principal),
-                }
-            }
-        }
-    }
-}
-
-impl From<ValidationError> for RequestError {
-    fn from(err: ValidationError) -> RequestError {
-        match err {
-            ValidationError::RecordValidationError(err) => err.into(),
-            ValidationError::ExternalCanisterValidationError(err) => err.into(),
-        }
-    }
 }
 
 impl ModelValidator<RequestError> for Request {
